Rename to tensor_augmentations
